Move API to namespace. Update Rails.
