+++ app/controllers/api/v1/places_controller.rb
@@ -1,0 +1,2 @@
+class API::V1::PlacesController < API::V1::APIController
+end
