+++ app/controllers/api/v1/profiles_controller.rb
@@ -1,0 +1,2 @@
+class API::V1::ProfilesController < API::V1::APIController
+end
