+++ app/controllers/api/v1/reports_controller.rb
@@ -1,0 +1,2 @@
+class API::V1::ReportsController < API::V1::APIController
+end
