+++ app/controllers/api/v1/users_controller.rb
@@ -1,0 +1,2 @@
+class API::V1::UsersController < API::V1::APIController
+end
