--- app/controllers/profiles_controller.rb
@@ -1,2 +1,0 @@
-class ProfilesController < ApplicationController
-end
