--- app/controllers/users_controller.rb
@@ -1,2 +1,0 @@
-class UsersController < ApplicationController
-end
