--- app/resources/aggregator_resource.rb
@@ -1,3 +1,0 @@
-class AggregatorResource < JSONAPI::Resource
-  attributes :name, :modifier, :description, :return_type
-end
