+++ app/resources/api/v1/aggregator_resource.rb
@@ -1,0 +1,9 @@
+module API
+  module V1
+    class AggregatorResource < JSONAPI::Resource
+
+      attributes :name, :modifier, :description, :return_type
+
+    end
+  end
+end
