+++ app/resources/api/v1/data_collection_resource.rb
@@ -1,0 +1,15 @@
+module API
+  module V1
+    class DataCollectionResource < JSONAPI::Resource
+
+      attribute :title
+      has_many :data_points
+
+      def ordered_data_points
+        """When DataPoints are included,
+          should show data points with order attached"""
+      end
+
+    end
+  end
+end
