+++ config/routes.rb
@@ -1,11 +1,18 @@
+require 'api_version'
+require 'subdomain_constraint'
+
 Rails.application.routes.draw do
 
-  jsonapi_resources :profiles, except: [:delete]
-  jsonapi_resources :places,   except: [:delete]
-  jsonapi_resources :reports,  except: [:delete]
-  jsonapi_resources :users,    only: [:show, :create]
+  namespace :api, constraints: SubdomainConstraint.new(/^api/), path: '' do
+    api_version(APIVersion.new(version: 1, default: true).params) do
+      jsonapi_resources :profiles, except: [:delete]
+      jsonapi_resources :places,   except: [:delete]
+      jsonapi_resources :reports,  except: [:delete]
+      jsonapi_resources :users,    only: [:show, :create]
 
-  post '/users/sign_in' => 'sessions#create'
+      post '/users/sign_in' => 'sessions#create'
+    end
+  end
 
   # TODO: Add a root path.
