+++ lib/subdomain_constraint.rb
@@ -1,0 +1,9 @@
+class SubdomainConstraint
+  def initialize(subdomain)
+    @subdomain = subdomain
+  end
+
+  def matches?(request)
+    @subdomain.match request.host
+  end
+end
